Move TypographyPreview to prod so that the preview can be rendered in Android Studio.
